django_react_02/backend/requirements.txt
django-environ==0.4.5
django-extensions==2.1.0
django-test-plus==1.1.1
django-url-checks==0.1.0
Django>=2.1,<2.2
djangorestframework==3.8.2
djangorestframework-xml>=1.3.0
defusedxml>=0.5.0
factory_boy==2.11.1
Faker==0.9.2
flake8==3.5.0
flake8-bugbear==18.2.0
flake8-docstrings==1.3.0
ipython==6.4.0
jupyter==1.0.0
pep8-naming==0.7.0
pre-commit==1.10.4
psycopg2-binary
pytz==2018.5
ShopifyAPI>=5.1.2